+++ packages/web3/tsconfig.eslint.json
@@ -1,4 +1,7 @@
 {
     "extends": "./tsconfig.json",
-    "include": ["src/**/*"]
+    "compilerOptions": {
+        "types": ["vitest/globals"]
+    },
+    "include": ["src/**/*", "tests/**/*"]
 }
