+++ packages/web3/tsconfig.json
@@ -2,8 +2,7 @@
     "extends": "./../tsconfig.base.json",
     "compilerOptions": {
         "baseUrl": "./src" /* Specify the root folder within your source files. */,
-        "outDir": "./dist" /* Specify an output folder for all emitted files. */,
-        "types": ["vitest/globals"]
+        "outDir": "./dist" /* Specify an output folder for all emitted files. */
     },
     "include": ["src/**/*"]
 }
